# CUDA: MMQ code deduplication + iquant support