… ggml/src/ggml-cuda/mmq.cuh
… ggml/src/ggml-cuda/template-instances/generate_cu_files.py
… ggml/src/ggml-cuda/template-instances/mmq-instance-iq1_s.cu
… ggml/src/ggml-cuda/template-instances/mmq-instance-iq2_s.cu
… ggml/src/ggml-cuda/template-instances/mmq-instance-iq2_xs.cu
… ggml/src/ggml-cuda/template-instances/mmq-instance-iq2_xxs.cu
… ggml/src/ggml-cuda/template-instances/mmq-instance-iq3_s.cu
… ggml/src/ggml-cuda/template-instances/mmq-instance-iq3_xxs.cu