This PR deduplicates the MMQ CUDA code and adds support for all i-quants other than qi1_m. The deduplication is done by converting the data to q8_0 or q8_1 with a block size of 32 or 16 upon loading. For the int8 tensor core kernels this has turned out to be faster. For q6_K and for the kernels using `__dp4a` this was slower so I left those unchanged. All i-quants other than qi1_m can then be supported by simply writing functions that load the data and convert it to q8. For qi1_m there is no support and frankly I don't think it would be worthwhile to add since the quality degradation of that format is very high.