@JohannesGaessler 2024-07-15
<details>
<summary>Performance</summary>

| Model                           | GPU        |     Microbatch size | Test   | t/s master |   t/s PR | Speedup |
| :------------------------------ | :--------- | ------------------: | :----- | ---------: | -------: | ------: |
| llama 8B IQ1_S - 1.5625 bpw     | RX 6800    |                  16 | pp2048 |      71.36 |   245.83 |    3.45 |
| llama 8B IQ1_S - 1.5625 bpw     | RX 6800    |                  32 | pp2048 |     126.88 |   341.29 |    2.69 |
| llama 8B IQ1_S - 1.5625 bpw     | RX 6800    |                  64 | pp2048 |     197.06 |   394.82 |    2.00 |
| llama 8B IQ1_S - 1.5625 bpw     | RX 6800    |                 128 | pp2048 |     344.40 |   495.76 |    1.44 |
| llama 8B IQ1_S - 1.5625 bpw     | RX 6800    |                 256 | pp2048 |     527.19 |   590.11 |    1.12 |
| llama 8B IQ1_S - 1.5625 bpw     | RX 6800    |                 512 | pp2048 |     579.49 |   598.22 |    1.03 |
| llama 8B IQ1_S - 1.5625 bpw     | RX 6800    |                1024 | pp2048 |     590.47 |   669.40 |    1.13 |
| llama 8B IQ1_S - 1.5625 bpw     | RX 6800    |                2048 | pp2048 |     578.68 |   613.85 |    1.06 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 3090   |                  16 | pp2048 |     355.54 |  1208.00 |    3.40 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 3090   |                  32 | pp2048 |     669.66 |  1916.56 |    2.86 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 3090   |                  64 | pp2048 |    1272.18 |  2760.97 |    2.17 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 3090   |                 128 | pp2048 |    2115.73 |  3326.07 |    1.57 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 3090   |                 256 | pp2048 |    3060.69 |  3768.10 |    1.23 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 3090   |                 512 | pp2048 |    3597.01 |  3938.13 |    1.09 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 3090   |                1024 | pp2048 |    4169.33 |  4040.79 |    0.97 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 3090   |                2048 | pp2048 |    4194.84 |  3933.14 |    0.94 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 4090   |                  16 | pp2048 |     537.90 |  1956.16 |    3.64 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 4090   |                  32 | pp2048 |    1057.79 |  4015.56 |    3.80 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 4090   |                  64 | pp2048 |    1962.03 |  6245.68 |    3.18 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 4090   |                 128 | pp2048 |    3690.13 |  8027.31 |    2.18 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 4090   |                 256 | pp2048 |    5947.65 | 10379.01 |    1.75 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 4090   |                 512 | pp2048 |    7911.42 | 11477.14 |    1.45 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 4090   |                1024 | pp2048 |    9016.49 | 11101.78 |    1.23 |
| llama 8B IQ1_S - 1.5625 bpw     | RTX 4090   |                2048 | pp2048 |    8936.54 |  9919.53 |    1.11 |
| llama 8B IQ1_S - 1.5625 bpw     | P40        |                  16 | pp2048 |      37.65 |   389.08 |   10.33 |
| llama 8B IQ1_S - 1.5625 bpw     | P40        |                  32 | pp2048 |      74.28 |   547.24 |    7.37 |
| llama 8B IQ1_S - 1.5625 bpw     | P40        |                  64 | pp2048 |     117.27 |   616.19 |    5.25 |
| llama 8B IQ1_S - 1.5625 bpw     | P40        |                 128 | pp2048 |     188.05 |   719.08 |    3.82 |
| llama 8B IQ1_S - 1.5625 bpw     | P40        |                 256 | pp2048 |     295.37 |   810.74 |    2.74 |
| llama 8B IQ1_S - 1.5625 bpw     | P40        |                 512 | pp2048 |     415.42 |   850.97 |    2.05 |
| llama 8B IQ1_S - 1.5625 bpw     | P40        |                1024 | pp2048 |     495.29 |   837.45 |    1.69 |
| llama 8B IQ1_S - 1.5625 bpw     | P40        |                2048 | pp2048 |     522.94 |   800.22 |    1.53 |
| llama 8B IQ2_S - 2.5 bpw        | RX 6800    |                  16 | pp2048 |      70.50 |   212.31 |    3.01 |
| llama 8B IQ2_S - 2.5 bpw        | RX 6800    |                  32 | pp2048 |     125.24 |   310.58 |    2.48 |
| llama 8B IQ2_S - 2.5 bpw        | RX 6800    |                  64 | pp2048 |     195.56 |   379.75 |    1.94 |
| llama 8B IQ2_S - 2.5 bpw        | RX 6800    |                 128 | pp2048 |     347.37 |   479.09 |    1.38 |
| llama 8B IQ2_S - 2.5 bpw        | RX 6800    |                 256 | pp2048 |     533.66 |   565.46 |    1.06 |
| llama 8B IQ2_S - 2.5 bpw        | RX 6800    |                 512 | pp2048 |     585.16 |   572.06 |    0.98 |
| llama 8B IQ2_S - 2.5 bpw        | RX 6800    |                1024 | pp2048 |     595.01 |   635.16 |    1.07 |
| llama 8B IQ2_S - 2.5 bpw        | RX 6800    |                2048 | pp2048 |     579.77 |   581.30 |    1.00 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 3090   |                  16 | pp2048 |     322.96 |  1110.28 |    3.44 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 3090   |                  32 | pp2048 |     611.47 |  1727.85 |    2.83 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 3090   |                  64 | pp2048 |    1170.72 |  2544.22 |    2.17 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 3090   |                 128 | pp2048 |    1991.36 |  3084.78 |    1.55 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 3090   |                 256 | pp2048 |    2954.34 |  3477.51 |    1.18 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 3090   |                 512 | pp2048 |    3505.89 |  3657.13 |    1.04 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 3090   |                1024 | pp2048 |    4091.96 |  3671.58 |    0.90 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 3090   |                2048 | pp2048 |    4143.53 |  3527.68 |    0.85 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 4090   |                  16 | pp2048 |     515.00 |  2145.23 |    4.17 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 4090   |                  32 | pp2048 |    1012.26 |  3573.84 |    3.53 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 4090   |                  64 | pp2048 |    1888.40 |  5693.41 |    3.01 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 4090   |                 128 | pp2048 |    3572.62 |  7480.11 |    2.09 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 4090   |                 256 | pp2048 |    5830.42 |  9519.30 |    1.63 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 4090   |                 512 | pp2048 |    7597.60 |  9945.81 |    1.31 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 4090   |                1024 | pp2048 |    8510.08 |  9521.63 |    1.12 |
| llama 8B IQ2_S - 2.5 bpw        | RTX 4090   |                2048 | pp2048 |    7950.37 |  8228.50 |    1.03 |
| llama 8B IQ2_S - 2.5 bpw        | P40        |                  16 | pp2048 |      36.51 |   348.53 |    9.55 |
| llama 8B IQ2_S - 2.5 bpw        | P40        |                  32 | pp2048 |      72.10 |   475.48 |    6.59 |
| llama 8B IQ2_S - 2.5 bpw        | P40        |                  64 | pp2048 |     113.47 |   587.33 |    5.18 |
| llama 8B IQ2_S - 2.5 bpw        | P40        |                 128 | pp2048 |     179.35 |   699.83 |    3.90 |
| llama 8B IQ2_S - 2.5 bpw        | P40        |                 256 | pp2048 |     285.87 |   790.54 |    2.77 |
| llama 8B IQ2_S - 2.5 bpw        | P40        |                 512 | pp2048 |     406.92 |   830.63 |    2.04 |
| llama 8B IQ2_S - 2.5 bpw        | P40        |                1024 | pp2048 |     486.02 |   819.36 |    1.69 |
| llama 8B IQ2_S - 2.5 bpw        | P40        |                2048 | pp2048 |     511.33 |   781.25 |    1.53 |
| llama 8B IQ2_XS - 2.3125 bpw    | RX 6800    |                  16 | pp2048 |      71.09 |   211.20 |    2.97 |
| llama 8B IQ2_XS - 2.3125 bpw    | RX 6800    |                  32 | pp2048 |     126.50 |   307.75 |    2.43 |
| llama 8B IQ2_XS - 2.3125 bpw    | RX 6800    |                  64 | pp2048 |     196.70 |   372.89 |    1.90 |
| llama 8B IQ2_XS - 2.3125 bpw    | RX 6800    |                 128 | pp2048 |     344.00 |   467.61 |    1.36 |
| llama 8B IQ2_XS - 2.3125 bpw    | RX 6800    |                 256 | pp2048 |     526.33 |   553.45 |    1.05 |
| llama 8B IQ2_XS - 2.3125 bpw    | RX 6800    |                 512 | pp2048 |     579.11 |   563.73 |    0.97 |
| llama 8B IQ2_XS - 2.3125 bpw    | RX 6800    |                1024 | pp2048 |     589.49 |   625.53 |    1.06 |
| llama 8B IQ2_XS - 2.3125 bpw    | RX 6800    |                2048 | pp2048 |     578.51 |   575.84 |    1.00 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 3090   |                  16 | pp2048 |     330.14 |  1128.10 |    3.42 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 3090   |                  32 | pp2048 |     621.15 |  1729.91 |    2.78 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 3090   |                  64 | pp2048 |    1186.35 |  2510.04 |    2.12 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 3090   |                 128 | pp2048 |    1982.17 |  2999.54 |    1.51 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 3090   |                 256 | pp2048 |    2937.32 |  3381.81 |    1.15 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 3090   |                 512 | pp2048 |    3486.24 |  3555.15 |    1.02 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 3090   |                1024 | pp2048 |    3994.26 |  3628.69 |    0.91 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 3090   |                2048 | pp2048 |    4110.74 |  3545.34 |    0.86 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 4090   |                  16 | pp2048 |     529.43 |  2231.90 |    4.22 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 4090   |                  32 | pp2048 |    1042.73 |  3674.42 |    3.52 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 4090   |                  64 | pp2048 |    1935.80 |  5707.91 |    2.95 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 4090   |                 128 | pp2048 |    3631.85 |  7439.87 |    2.05 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 4090   |                 256 | pp2048 |    5886.87 |  9381.76 |    1.59 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 4090   |                 512 | pp2048 |    7850.31 | 10109.82 |    1.29 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 4090   |                1024 | pp2048 |    8984.33 |  9941.22 |    1.11 |
| llama 8B IQ2_XS - 2.3125 bpw    | RTX 4090   |                2048 | pp2048 |    8908.32 |  8988.81 |    1.01 |
| llama 8B IQ2_XS - 2.3125 bpw    | P40        |                  16 | pp2048 |      37.62 |   349.74 |    9.30 |
| llama 8B IQ2_XS - 2.3125 bpw    | P40        |                  32 | pp2048 |      74.33 |   464.01 |    6.24 |
| llama 8B IQ2_XS - 2.3125 bpw    | P40        |                  64 | pp2048 |     117.46 |   592.29 |    5.04 |
| llama 8B IQ2_XS - 2.3125 bpw    | P40        |                 128 | pp2048 |     189.35 |   701.98 |    3.71 |
| llama 8B IQ2_XS - 2.3125 bpw    | P40        |                 256 | pp2048 |     296.17 |   791.88 |    2.67 |
| llama 8B IQ2_XS - 2.3125 bpw    | P40        |                 512 | pp2048 |     415.48 |   833.32 |    2.01 |
| llama 8B IQ2_XS - 2.3125 bpw    | P40        |                1024 | pp2048 |     495.55 |   823.80 |    1.66 |
| llama 8B IQ2_XS - 2.3125 bpw    | P40        |                2048 | pp2048 |     521.07 |   788.80 |    1.51 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RX 6800    |                  16 | pp2048 |      71.13 |   210.94 |    2.97 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RX 6800    |                  32 | pp2048 |     126.59 |   310.59 |    2.45 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RX 6800    |                  64 | pp2048 |     196.74 |   386.29 |    1.96 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RX 6800    |                 128 | pp2048 |     343.90 |   486.67 |    1.42 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RX 6800    |                 256 | pp2048 |     526.43 |   579.28 |    1.10 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RX 6800    |                 512 | pp2048 |     579.59 |   586.98 |    1.01 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RX 6800    |                1024 | pp2048 |     592.60 |   656.27 |    1.11 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RX 6800    |                2048 | pp2048 |     580.42 |   601.83 |    1.04 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 3090   |                  16 | pp2048 |     345.83 |  1140.41 |    3.30 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 3090   |                  32 | pp2048 |     642.45 |  1856.45 |    2.89 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 3090   |                  64 | pp2048 |    1224.86 |  2797.71 |    2.28 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 3090   |                 128 | pp2048 |    2055.07 |  3488.32 |    1.70 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 3090   |                 256 | pp2048 |    2989.49 |  3936.71 |    1.32 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 3090   |                 512 | pp2048 |    3516.58 |  4161.70 |    1.18 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 3090   |                1024 | pp2048 |    4078.77 |  4278.76 |    1.05 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 3090   |                2048 | pp2048 |    4144.55 |  4168.99 |    1.01 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 4090   |                  16 | pp2048 |     532.63 |  2266.50 |    4.26 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 4090   |                  32 | pp2048 |    1045.01 |  3885.14 |    3.72 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 4090   |                  64 | pp2048 |    1944.40 |  6300.56 |    3.24 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 4090   |                 128 | pp2048 |    3657.66 |  8300.52 |    2.27 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 4090   |                 256 | pp2048 |    5910.69 | 10773.95 |    1.82 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 4090   |                 512 | pp2048 |    7885.28 | 11858.44 |    1.50 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 4090   |                1024 | pp2048 |    8996.48 | 11471.93 |    1.28 |
| llama 8B IQ2_XXS - 2.0625 bpw   | RTX 4090   |                2048 | pp2048 |    8946.50 | 10195.18 |    1.14 |
| llama 8B IQ2_XXS - 2.0625 bpw   | P40        |                  16 | pp2048 |      37.49 |   363.76 |    9.70 |
| llama 8B IQ2_XXS - 2.0625 bpw   | P40        |                  32 | pp2048 |      73.96 |   534.20 |    7.22 |
| llama 8B IQ2_XXS - 2.0625 bpw   | P40        |                  64 | pp2048 |     116.84 |   592.01 |    5.07 |
| llama 8B IQ2_XXS - 2.0625 bpw   | P40        |                 128 | pp2048 |     188.23 |   701.77 |    3.73 |
| llama 8B IQ2_XXS - 2.0625 bpw   | P40        |                 256 | pp2048 |     295.12 |   792.59 |    2.69 |
| llama 8B IQ2_XXS - 2.0625 bpw   | P40        |                 512 | pp2048 |     414.91 |   834.12 |    2.01 |
| llama 8B IQ2_XXS - 2.0625 bpw   | P40        |                1024 | pp2048 |     494.95 |   824.74 |    1.67 |
| llama 8B IQ2_XXS - 2.0625 bpw   | P40        |                2048 | pp2048 |     520.41 |   789.70 |    1.52 |
| llama 8B IQ3_S - 3.4375 bpw     | RX 6800    |                  16 | pp2048 |      70.10 |   211.71 |    3.02 |
| llama 8B IQ3_S - 3.4375 bpw     | RX 6800    |                  32 | pp2048 |     124.91 |   314.11 |    2.51 |
| llama 8B IQ3_S - 3.4375 bpw     | RX 6800    |                  64 | pp2048 |     195.43 |   391.09 |    2.00 |
| llama 8B IQ3_S - 3.4375 bpw     | RX 6800    |                 128 | pp2048 |     346.08 |   493.11 |    1.42 |
| llama 8B IQ3_S - 3.4375 bpw     | RX 6800    |                 256 | pp2048 |     530.73 |   584.64 |    1.10 |
| llama 8B IQ3_S - 3.4375 bpw     | RX 6800    |                 512 | pp2048 |     584.42 |   590.85 |    1.01 |
| llama 8B IQ3_S - 3.4375 bpw     | RX 6800    |                1024 | pp2048 |     590.78 |   657.73 |    1.11 |
| llama 8B IQ3_S - 3.4375 bpw     | RX 6800    |                2048 | pp2048 |     581.40 |   601.84 |    1.04 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 3090   |                  16 | pp2048 |     328.90 |  1003.01 |    3.05 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 3090   |                  32 | pp2048 |     625.84 |  1685.31 |    2.69 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 3090   |                  64 | pp2048 |    1198.81 |  2729.18 |    2.28 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 3090   |                 128 | pp2048 |    2030.01 |  3472.59 |    1.71 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 3090   |                 256 | pp2048 |    2994.96 |  3966.49 |    1.32 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 3090   |                 512 | pp2048 |    3525.21 |  4151.76 |    1.18 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 3090   |                1024 | pp2048 |    4083.91 |  4201.95 |    1.03 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 3090   |                2048 | pp2048 |    4072.81 |  4026.83 |    0.99 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 4090   |                  16 | pp2048 |     501.48 |  1707.30 |    3.40 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 4090   |                  32 | pp2048 |     989.87 |  3069.61 |    3.10 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 4090   |                  64 | pp2048 |    1846.74 |  5703.12 |    3.09 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 4090   |                 128 | pp2048 |    3479.72 |  7787.52 |    2.24 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 4090   |                 256 | pp2048 |    5721.47 | 10473.88 |    1.83 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 4090   |                 512 | pp2048 |    7494.42 | 11184.52 |    1.49 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 4090   |                1024 | pp2048 |    8488.27 | 10522.71 |    1.24 |
| llama 8B IQ3_S - 3.4375 bpw     | RTX 4090   |                2048 | pp2048 |    7954.46 |  8944.74 |    1.12 |
| llama 8B IQ3_S - 3.4375 bpw     | P40        |                  16 | pp2048 |      36.40 |   335.57 |    9.22 |
| llama 8B IQ3_S - 3.4375 bpw     | P40        |                  32 | pp2048 |      71.98 |   521.85 |    7.25 |
| llama 8B IQ3_S - 3.4375 bpw     | P40        |                  64 | pp2048 |     112.97 |   575.59 |    5.10 |
| llama 8B IQ3_S - 3.4375 bpw     | P40        |                 128 | pp2048 |     180.55 |   688.69 |    3.81 |
| llama 8B IQ3_S - 3.4375 bpw     | P40        |                 256 | pp2048 |     285.55 |   780.64 |    2.73 |
| llama 8B IQ3_S - 3.4375 bpw     | P40        |                 512 | pp2048 |     406.24 |   822.90 |    2.03 |
| llama 8B IQ3_S - 3.4375 bpw     | P40        |                1024 | pp2048 |     486.12 |   810.37 |    1.67 |
| llama 8B IQ3_S - 3.4375 bpw     | P40        |                2048 | pp2048 |     510.18 |   773.40 |    1.52 |
| llama 8B IQ3_S mix - 3.66 bpw   | RX 6800    |                  16 | pp2048 |      73.51 |   215.02 |    2.93 |
| llama 8B IQ3_S mix - 3.66 bpw   | RX 6800    |                  32 | pp2048 |     130.61 |   311.82 |    2.39 |
| llama 8B IQ3_S mix - 3.66 bpw   | RX 6800    |                  64 | pp2048 |     203.35 |   382.47 |    1.88 |
| llama 8B IQ3_S mix - 3.66 bpw   | RX 6800    |                 128 | pp2048 |     346.54 |   481.26 |    1.39 |
| llama 8B IQ3_S mix - 3.66 bpw   | RX 6800    |                 256 | pp2048 |     528.16 |   571.78 |    1.08 |
| llama 8B IQ3_S mix - 3.66 bpw   | RX 6800    |                 512 | pp2048 |     582.26 |   580.56 |    1.00 |
| llama 8B IQ3_S mix - 3.66 bpw   | RX 6800    |                1024 | pp2048 |     601.59 |   646.13 |    1.07 |
| llama 8B IQ3_S mix - 3.66 bpw   | RX 6800    |                2048 | pp2048 |     579.66 |   591.26 |    1.02 |
| llama 8B IQ3_S mix - 3.66 bpw   | RTX 3090   |                  16 | pp2048 |     361.53 |  1031.70 |    2.85 |
| llama 8B IQ3_S mix - 3.66 bpw   | RTX 3090   |                  32 | pp2048 |     677.51 |  1720.06 |    2.54 |
| llama 8B IQ3_S mix - 3.66 bpw   | RTX 3090   |                  64 | pp2048 |    1275.19 |  2740.30 |    2.15 |
…